requirements.txt
fastapi==0.103.0
uvicorn[standard]==0.22.0
sqlalchemy==2.0.20
aiosqlite==0.18.0
openai
faiss-cpu==1.7.4
python-dotenv==1.0.0
numpy==1.26.4
